chore(Poetry): Downgrade from 1.2.0a2 to 1.1.11
We previously used a pre-released version of Poetry for poetry lock
--check, which validates that pyproject.toml and poetry.lock are kept in
sync, preventing us from forgetting to run poetry install. With
pre-commit, however, this new command is not necessary. We can
simply run poetry lock --no-update, and if it modifies the lock file,
pre-commit will consider the hook to have failed. Thus, we can fall back
to the latest stable version of Poetry. Poetry 1.2 is presently
incompatible with asdf, which we intend to use.