fix(action): Only run dockerd-rootless.sh if told
GitHub Actions made an unannounced change impacting how we determine
whether the rootless Docker install script started the Docker daemon.
Previously, "API listen on $XDG_RUNTIME_DIR/docker.sock" was present in
the script's output if the Docker daemon was started, but this is no
longer guaranteed. Running dockerd-rootless.sh when the Docker daemon
has already been started results in an error, so only run it when the
install script outputs an explicit instruction to do so.